sync adapter

+++ app/src/main/java/com/example/rishabh/curotest/BodyRequest/LogValuesData.java
@@ -9,7 +9,7 @@ public class LogValuesData {
   public Integer tasktemplate_id;
   public String date_time;
   public String logged_time;
-  public Long value;
+  public double value;
   public Integer timeslot_id;
   public Integer vitaldataattribute_id;
   public Integer server_id;
